Minor cleaning
