+++ src/ModelOrderReduction/component/forcefield/HyperReducedTriangleFEMForceField.inl
@@ -14,10 +14,9 @@
 *                                                                             *
 * Contact information: https://project.inria.fr/modelorderreduction/contact   *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_FORCEFIELD_HYPERREDUCEDTRIANGLEFEMFORCEFIELD_INL
-#define SOFA_COMPONENT_FORCEFIELD_HYPERREDUCEDTRIANGLEFEMFORCEFIELD_INL
+#pragma once
 
-#include "HyperReducedTriangleFEMForceField.h"
+#include <ModelOrderReduction/component/forcefield/HyperReducedTriangleFEMForceField.h>
 #include <sofa/core/visual/VisualParams.h>
 #include <sofa/core/MechanicalParams.h>
 #include <sofa/core/ObjectFactory.h>
@@ -27,19 +26,9 @@
 #include <iostream> //for debugging
 #include <vector>
 #include <sofa/defaulttype/VecTypes.h>
-#include "../loader/MatrixLoader.h"
+#include <ModelOrderReduction/component/loader/MatrixLoader.h>
 
-//#include "config.h"
-
-// #define DEBUG_TRIANGLEFEM
-
-namespace sofa
-{
-
-namespace component
-{
-
-namespace forcefield
+namespace sofa::component::forcefield
 {
 
 using sofa::component::loader::MatrixLoader;
@@ -409,12 +398,4 @@ void HyperReducedTriangleFEMForceField<DataTypes>::addKToMatrix(sofa::linearalge
         }
     }
 }
-
-
-} // namespace forcefield
-
-} // namespace component
-
-} // namespace sofa
-
-#endif // #ifndef SOFA_COMPONENT_FORCEFIELD_HYPERREDUCEDTRIANGLEFEMFORCEFIELD_INL
+} // namespace sofa::component::forcefield
